Trait dex_general_rpc::DexGeneralApiServer

pub trait DexGeneralApiServer<BlockHash, AccountId, AssetId>: Sized + Send + Sync + 'static {
    // Required methods
    fn get_pair_by_asset_id(
        &self,
        asset_0: AssetId,
        asset_1: AssetId,
        at: Option<BlockHash>
    ) -> RpcResult<Option<PairInfo<AccountId, NumberOrHex, AssetId>>>;
    fn get_amount_in_price(
        &self,
        supply: AssetBalance,
        path: Vec<AssetId>,
        at: Option<BlockHash>
    ) -> RpcResult<NumberOrHex>;
    fn get_amount_out_price(
        &self,
        supply: AssetBalance,
        path: Vec<AssetId>,
        at: Option<BlockHash>
    ) -> RpcResult<NumberOrHex>;
    fn get_estimate_lptoken(
        &self,
        asset_0: AssetId,
        asset_1: AssetId,
        amount_0_desired: AssetBalance,
        amount_1_desired: AssetBalance,
        amount_0_min: AssetBalance,
        amount_1_min: AssetBalance,
        at: Option<BlockHash>
    ) -> RpcResult<NumberOrHex>;
    fn calculate_remove_liquidity(
        &self,
        asset_0: AssetId,
        asset_1: AssetId,
        amount: AssetBalance,
        at: Option<BlockHash>
    ) -> RpcResult<Option<(AssetBalance, AssetBalance)>>;

    // Provided method
    fn into_rpc(self) -> RpcModule<Self>
       where BlockHash: Send + Sync + 'static + DeserializeOwned,
             AccountId: Send + Sync + 'static + Serialize,
             AssetId: Send + Sync + 'static + DeserializeOwned + Serialize { ... }
}

Server trait implementation for the DexGeneralApi RPC API.

Required Methods

fn get_pair_by_asset_id( &self, asset_0: AssetId, asset_1: AssetId, at: Option<BlockHash> ) -> RpcResult<Option<PairInfo<AccountId, NumberOrHex, AssetId>>>

fn get_amount_in_price( &self, supply: AssetBalance, path: Vec<AssetId>, at: Option<BlockHash> ) -> RpcResult<NumberOrHex>

fn get_amount_out_price( &self, supply: AssetBalance, path: Vec<AssetId>, at: Option<BlockHash> ) -> RpcResult<NumberOrHex>

fn get_estimate_lptoken( &self, asset_0: AssetId, asset_1: AssetId, amount_0_desired: AssetBalance, amount_1_desired: AssetBalance, amount_0_min: AssetBalance, amount_1_min: AssetBalance, at: Option<BlockHash> ) -> RpcResult<NumberOrHex>

fn calculate_remove_liquidity( &self, asset_0: AssetId, asset_1: AssetId, amount: AssetBalance, at: Option<BlockHash> ) -> RpcResult<Option<(AssetBalance, AssetBalance)>>

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where BlockHash: Send + Sync + 'static + DeserializeOwned, AccountId: Send + Sync + 'static + Serialize, AssetId: Send + Sync + 'static + DeserializeOwned + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.